+++ swim_tests/test/bool_wire_codegen.py
@@ -1,0 +1,10 @@
+# top=test_bool_wire_codegen::harness
+
+from cocotb import cocotb
+
+@cocotb.test()
+async def test(dut):
+    # NOTE: This test checks for miscompilation of the generated verilog,
+    # so no tests are actually performed
+    pass
+